# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Personal Finances",
    "description": "API from personal finances system",
    "version": "1.0"
  },
  "definitions": {
    "user": {
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "userEmail": {
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "transaction": {
      "properties": {
        "type": {
          "type": "string"
        },
        "value": {
          "type": "number"
        },
        "payd": {
          "type": "boolean"
        },
        "dateTransaction": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "user_id": {
          "type": "integer"
        }
      }
    }
  },
  "tags": [
    {
      "name": "Users",
      "description": "Users in system"
    },
    {
      "name": "Transactions",
      "description": "Transactions in system"
    }
  ],
  "host": "localhost:3333",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a user",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User that we what to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "New user is created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List all Users.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List one User by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Id from user",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update one User by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Id from user",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/user"
            },
            "description": "Data from update user",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update one User email by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Id from user",
            "required": true
          },
          {
            "name": "email",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/userEmail"
            },
            "description": "Email from user",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete one User by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Id from user",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/transaction/": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Create a transaction",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Transaction that we what to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/transaction"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "New Transaction is created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "List all Transactions.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/transaction/{id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "List one Transaction by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Id from transaction",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Transactions"
        ],
        "summary": "Update one Transaction by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Id from transaction",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/transaction"
            },
            "description": "Data from update transaction",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "summary": "Delete one Transaction by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Id from transaction",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}